requirements.txt
fastapi~=0.115.12
pydantic~=2.11.4
python-multipart~=0.0.9
uvicorn~=0.28.0
sqlalchemy~=2.0.24
pymysql~=1.1.0
cryptography~=41.0.4
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4
httpx~=0.27.0
aiofiles~=24.1.0
pillow~=11.2.1
python-dotenv~=1.1.0